cli/pyproject.toml
[project]
name = "triad-cli"
version = "0.1.0"
description = "TRIAD — Multi-Agent Consensus System CLI"
requires-python = ">=3.11"
dependencies = [
    "typer[all]>=0.12.0",
    "rich>=13.7.0",
    "httpx>=0.27.0",
    "websockets>=13.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
triad = "triad_cli.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
